requirements.txt
absl-py==2.1.0
apache-airflow==2.7.2
cmake==3.31.4
configparser==7.1.0
ninja==1.11.1.3
numpy==2.2.2
pandas==2.2.3
pendulum==3.0.0
praw==7.8.1
python-dateutil==2.9.0.post0
setuptools==75.8.0
six==1.17.0
time-machine==2.16.0
tzdata==2025.1
wheel==0.45.1
